+++ led.c
@@ -3,10 +3,10 @@
 #include "led.h"
 #include "uart.h"
 
-extern unsigned int time;
+extern unsigned long time;
 extern struct led_state_t led_state;
 
-char time_exceed(unsigned int time_pre, unsigned int cycle_ms) {
+char time_exceed(unsigned long time_pre, unsigned long cycle_ms) {
     return (char)((time - time_pre) > (cycle_ms/10));
 }
 
@@ -15,7 +15,7 @@ void set_led_state(char state) {
 }
 
 void led_run() {
-    static int interval_time = 0;
+    static long interval_time = 0;
 
     if(led_state.state != led_state.pre_state) {
         led_state.pre_state = led_state.state;
